==> 03_extract_zetas.R <==
####################################################
# Mariona Roigé Valiente, 2016                     #
# Ecological Informatics, BPRC, Lincoln University #              			  
# mariona.roige.valiente@gmail.com       	   #
####################################################

setwd('/home/mariona/Documents/00_PhD/0_Thesis_outline/03_Zetadiversity/02_paperMME/Materials for online')
load('zetavalues.RData') # Import the workspace generated by script 01.
library(zetadiv)
library(plyr)
library(dplyr)
library(stringr)

ls()
summary(zetavalues) # zetavalues is a list of lists. 


coerce <- function(s) {					# Extract the parts of each element of zetavalues that contain
coerced <- as.data.frame(lapply(s[1:2], as.data.frame))
names(coerced) <- c('order','zeta') 			# the information we are interested in.
return(coerced)						# Function to coerce elements [1:2] of the inner lists into data.frames()
}

zetas_coerced <- lapply(zetavalues, coerce)

# Create new variable, length max dimension of Zi with column 2 of every data.frame() plus NA

zrefill <- function(x){				# Function to refill each data.frame in the list with NA to 
	zfill <- rep(NA,max_reg)		# homogenize their dimensions. 
      zfill[1:length(x[,2])] <- x[,2]
	return(zfill)
				}

zetas_table <- ldply(zetas_coerced, zrefill)       # Coerce all the zeta values in one data.frame
nametable <- rep('namehere', max_reg)              # Create names for the columns
for (i in 1:max_reg){				   #
nametable[i] <- paste('zeta', i, sep ='')}	   #
names(zetas_table) <- nametable			   
numlong <- seq(1:ncell)				   # Create names for the rows. I.e: the cell number they belong to
numericalpart <- numlong[-pos_to_delete]
nametable2 <- rep('namehere',length(numericalpart))
for(i in 1:length(numericalpart)){
nametable2[i] <- paste('cell',numericalpart[i],sep = '')}
row.names(zetas_table) <- nametable2


norm_zetas <- zetas_table/zetas_table[,1]	# Calculate the normalized zeta_i values by dividing by z1


save.image('extractedzetas.RData')
