==> 04_select_clusters.R <==
####################################################
# Mariona Roigé Valiente, 2016                     #
# Ecological Informatics, BPRC, Lincoln University #              			  
# mariona.roige.valiente@gmail.com       	   #
####################################################

setwd('/home/mariona/Documents/00_PhD/0_Thesis_outline/03_Zetadiversity/02_paperMME/Materials for online')
load('extractedzetas.RData') # Import the workspace generated by script '03_extract_zetas.R'
library(zetadiv)
library(plyr)
library(dplyr)
library(stringr)

# Low Uncertainty clusters #################################

x <- norm_zetas						# 1- Select those clusters that have high z2 and high z3-z5. 

for (i in 1:4){						# the criteria I used is to ditch always the 1st quartile. 

if (i==1){
threshold <- summary(x[,i+1])[2]
goodx <- x[x[,i+1]>threshold,]
}else{
x <- goodx
xna <- which(is.na(x[,i+1]))
threshold <- summary(x[,i+1])[2]
goodx <- x[-xna,]
goodx_t <- goodx[goodx[,i+1]>threshold,]
goodx <- goodx_t
}
}	

goodz <- goodx # contains the values of zeta (1 to max order) for the clusters
     	       # that have a high z2 and a high z3-z5

# Medium Uncertainty clusters ###############################

y <- norm_zetas

th_z2 <- summary(y$zeta2)[2]				# 2- Select those clusters with high z2 but not high z3-z5
goodz2 <- y[y$zeta2>th_z2,]  
mediumy <- goodz2[-(match(row.names(goodx),row.names(goodz2))),]

mediumz <- mediumy # contains the values of zeta(1 to max order) for the clusters
		   # that have a high z2 but not high z3-z5


# High Uncertainty clusters #################################

goodandmed <- c(row.names(mediumz),row.names(goodz))	# 3- Select those clusters with low z2 and low z3-z5
badz <- norm_zetas[-(match(goodandmed,row.names(norm_zetas))),]

# test: dim(badz)[1]+dim(mediumz)[1]+dim(goodz)[1] == dim(norm_zetas)[1]

save.image('selectedzetas.RData')
