==> 05_plot_zetas.R <==
####################################################
# Mariona Roigé Valiente, 2016                     #
# Ecological Informatics, BPRC, Lincoln University #              			  
# mariona.roige.valiente@gmail.com       	   #
####################################################

setwd('/home/mariona/Documents/00_PhD/0_Thesis_outline/03_Zetadiversity/02_paperMME/Materials for online')
load('extractedzetas.RData') # Import the workspace generated by script '03_extract_zetas.R'
library(zetadiv)
library(plyr)
library(dplyr)
library(stringr)
library(RColorBrewer) #to use brewer.pal
library(fields) #to use designer.colors

summary(zetas_table) # In this data.frame we stored all the values of zeta for all those non empty or non single neurons
summary(norm_zetas)  # In this data.frame we stored all the values of normalized zetas for all those non empty or non single neurons  

# Fill the empty neurons and the single cells with ROWS of NAs

sortedtofill <- sort(c(empty_cells,single_cells))

for (i in sortedtofill){		
toinsert <- rep(NA,max_reg)
zetas_table <- rbind(rbind(zetas_table[1:i-1,],toinsert),zetas_table[i:ncell,])
row.names(zetas_table)[i] <- paste('cell',i, sep='')
}; zetas_table <- zetas_table[-(dim(zetas_table)[1]),]


for (i in sortedtofill){		
toinsert <- rep(NA,max_reg)
norm_zetas<- rbind(rbind(norm_zetas[1:i-1,],toinsert),norm_zetas[i:ncell,])
row.names(norm_zetas)[i] <- paste('cell',i, sep='')
}; norm_zetas <- norm_zetas[-(dim(norm_zetas)[1]),]


#######################################################
# Plotting the zeta values in the map grids
#######################################################
# code for the map grids found in : http://stackoverflow.com/questions/19858729/r-package-kohonen-how-to-plot-hexagons-instead-of-circles-as-in-matlab-som-too. All credit to NBremer.  

# Data preparation. The plotting function works inversely to the 
# nomenclature of our cells. To be able to plot with this function
# we have to create a vector of equivalences between the name of the
# cells provided by Matlab and the SOM Hexagon plotting function in R. 

SOM_Rows <- 11	# This we have to define it provided the SOM output map produced in
SOM_Columns <- 9   # '01_script_matlab.m'. Note that (nrowsSOM*ncolsSOM==ncell) holds TRUE

positions <- seq(1:ncell)
chunks <- split(positions, ceiling(seq_along(positions)/SOM_Rows))	  # Create a vector of equivalence 
positionsSOM <- as.numeric(unlist(lapply(chunks,sort,decreasing = TRUE))) # between cell numbers and plot positions	
zetas_table_p <- data.frame(zetas_table,positionsSOM)
zetas_table_sorted <- zetas_table_p[order(zetas_table_p$positionsSOM,zetas_table_p$positionsSOM),] # Data frame to plot raw zetas
norm_zetas_p <- data.frame(norm_zetas, positionsSOM)
norm_zetas_sorted <- norm_zetas_p[order(norm_zetas_p$positionsSOM, norm_zetas_p$positionsSOM),]    # Data frame to plot normalized zetas

# Function Hexagon by NBremer. 

Hexagon <- function (x, y, unitcell = 1, col = col) {
  polygon(c(x, x, x + unitcell/2, x + unitcell, x + unitcell, 
            x + unitcell/2), c(y + unitcell * 0.125, y + unitcell * 
                               0.875, y + unitcell * 1.125, y + unitcell * 0.875, 
                               y + unitcell * 0.125, y - unitcell * 0.125), 
          col = col, border=NA)
}

# Plot any zeta (z1 to zmax_reg)
# Insert in 'zetas_table_sorted$zeta1' any column of zetas_table_sorted  or 
# of norm_zetas_sorted you want to plot.

x <- as.vector(matrix(zetas_table_sorted$zeta1, nrow = SOM_Rows, ncol = SOM_Columns))

par(mar = c(0.4, 2, 2, 7))
plot(0, 0, type = "n", axes = FALSE, xlim=c(0, SOM_Columns), 
     ylim=c(0, SOM_Rows), xlab="", ylab= "", asp=1)
ColRamp <- rev(designer.colors(n=50, col=brewer.pal(9, "Spectral")))
ColorCode <- rep("#FFFFFF", length(x)) #default is all white
Bins <- seq(min(x, na.rm=T), max(x, na.rm=T), length=length(ColRamp))
for (i in 1:length(x))
    if (!is.na(x[i])) ColorCode[i] <- ColRamp[which.min(abs(Bins-x[i]))] 
offset <- 0.5 #offset for the hexagons when moving up a row
for (row in 1:SOM_Rows) {
  for (column in 0:(SOM_Columns - 1)) 
     Hexagon(column + offset, row - 1, col = ColorCode[row + SOM_Rows * column])
  offset <- ifelse(offset, 0, 0.5)
}
image.plot(legend.only=TRUE, col=ColRamp, zlim=c(min(x, na.rm=T), max(x, na.rm=T)))

###################################################################
# Optional code to plot an empty SOM Map with each cell number    #
###################################################################

# The empty grid (colorless) with cell numbers

Labels <- seq(1:ncell)
Labels <- matrix (Labels , nrow = SOM_Rows, ncol = SOM_Columns)
Labels <- apply(Labels,2, sort, decreasing = T)

# Function Hexagon modified to plot text

Hexagon <- function (x, y, unitcell = 1, col = col) {
  polygon(c(x, x, x + unitcell/2, x + unitcell, x + unitcell,
            x + unitcell/2), c(y + unitcell * 0.125, y + unitcell *
                               0.875, y + unitcell * 1.125, y + unitcell * 0.875,
                               y + unitcell * 0.125, y - unitcell * 0.125),
          col = col, border='black' )
  text(x+0.5, y+0.5, labels = Labels[row,column+1], adj = NULL,
     pos = NULL, offset = 3, vfont = NULL,
     cex = 1, col = 'black', font = NULL)
}


x <- as.vector(matrix(seq(1:ncell), nrow = SOM_Rows, ncol = SOM_Columns))
par(mar = c(0.4, 2, 2, 5))

#png('gridnumbered.png') Uncomment for saving a *.png image
#pdf('gridnumbered.pdf') Uncomment for saving a *.pdf image

plot(0, 0, type = "n", axes = FALSE, xlim=c(0, SOM_Columns),
     ylim=c(0, SOM_Rows), xlab="", ylab= "", asp=1)
ColRamp <- rev(designer.colors(n=50, col=brewer.pal(9, "Spectral")))
ColorCode <- rep("#FFFFFF", length(x)) #default is all white
Bins <- seq(min(x, na.rm=T), max(x, na.rm=T), length=length(ColRamp))
for (i in 1:length(x))
    if (!is.na(x[i])) ColorCode[i] <- ColRamp[which.min(abs(Bins-x[i]))]
offset <- 0.5 #offset for the hexagons when moving up a row
for (row in 1:SOM_Rows) {
  for (column in 0:(SOM_Columns - 1))
     Hexagon(column + offset, row - 1, col = NULL) #null color
  offset <- ifelse(offset, 0, 0.5)
}
#dev.off() Uncomment if using png('*.png') or pdf('*.pdf')
